refactor(filename): change constants to nodeTypesMapping
