+++ src/app/pages/Studio/containers/Canvas/nodeTypesMapping.ts
